+++ app/scripts/options.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 // import Vue from 'vue/dist/vue.js';
 import Vuetify from 'vuetify';
 import Options from '../pages/options.vue'
-import "material-design-icons-iconfont/dist/material-design-icons.css"
+import '@mdi/font/css/materialdesignicons.css' // Ensure you are using css-loader
 import 'vuetify/dist/vuetify.css'
 import '../styles/options.css';
 
@@ -23,7 +23,9 @@ function initOptions() {
  * Initialize Vue
  */
 function initVue() {
-    Vue.use(Vuetify, {});
+    Vue.use(Vuetify, {
+        iconfont: 'mdi'
+    });
 
     new Vue({
         el: '#app',